hjwt authentication and authorization done

+++ backend/controllers/goalController.js
@@ -1,13 +1,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@desc get goals
 //@route GET /api/goals
 //@access Private
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -22,9 +23,10 @@ const setGoal = asyncHandler(async (req, res) => {
     }
      
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
-    res.status(200).json({message: 'set goal'})
+    res.status(200).json(goal)
 })
 
 //@desc UPDATE goals
@@ -38,6 +40,21 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    //make sure the logged in user mathces the goal user
+    if(goal.user.toString() !== user.id) {
+      res.status(401)
+      throw new Error('user not authorised')
+    }
+
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,
         req.body, {
@@ -58,6 +75,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('goal not found')
+    }
+
+     const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    //make sure the logged in user mathces the goal user
+    if(goal.user.toString() !== user.id) {
+      res.status(401)
+      throw new Error('user not authorised')
     }
 
     await goal.deleteOne()
